Fixed WriteMultiLevelPointer function

+++ MiniMem/Constants.cs
@@ -134,8 +134,8 @@ namespace MiniMem
 		public const int PROCESS_VM_READ = 0x0010;
 		public const int PROCESS_VM_WRITE = 0x0020;
 
-		public static int m_iNumberOfBytesRead;
-		public static int m_iNumberOfBytesWritten;
+		public static IntPtr m_iNumberOfBytesRead;
+		public static IntPtr m_iNumberOfBytesWritten;
 
 		public enum MessageType
 		{
